Fix fallback sandbox settings imports for flat execution
